Add algebra
Added a bunch of exposition on the
Kahler differentials

Also added a proof of the equivalence of
the first two wikipedia definitions of regularity.
The proof is way more elementary than I thought,
it doesn't require Krull's prinicpal ideal theorem.

Probably a few other odds and ends, I don't remember