one_mul working
Finally, after a lot of struggling, I have finally
successfully gotten a proof of
one_mul for the monoid.

The proof is horrendous--I know it can be golfed a lot,
and I also know that if I knew what I was doing the proof
would have taken a lot less time to write.
I'll be learning more about this as soon as I can.

For now, though, I've left in much of my scratch work
in the comments for posterity in this commit.
They are quite ugly, but perhaps someday it might
be useful to look at them?
May they rest in peace somewhere on a github server.

Also, add an example of morphisms using simp in a file.